+++ gamemodes/terrortown/gamemode/server/sv_player.lua
@@ -1209,8 +1209,8 @@ function GM:EntityTakeDamage(ent, dmginfo)
 	end
 
 	--Call a second Hook after Damage got processed, to output the real Damage that has been dealt
-	--Can you use return like this, so that hooks to "TTTPostEntityTakeDamage" could also return true or false to block the damage event afterwards
-	return hook.Run("TTTPostEntityTakeDamage", ent, dmginfo)
+	--Can you use return like this, so that hooks to "TTT2PostEntityTakeDamage" could also return true or false to block the damage event afterwards
+	return hook.Run("TTT2PostEntityTakeDamage", ent, dmginfo)
 end
 
 ---
@@ -1491,5 +1491,5 @@ end
 -- @hook
 -- @realm server
 -- @local
-function GM:TTTPostEntityTakeDamage(ent, dmginfo)
+function GM:TTT2PostEntityTakeDamage(ent, dmginfo)
 end